Fix (settings): Fixed base dir
When changing settings from a single file to a module, settings files are now a level deeper in the directory hierarchy. Production settings need this to be correctly set (it would still run fine on development settings). This would cause the application not to run on productions settings (on heroku). Now the base directory variable is correctly set to reflect the new directory hierarchy.