Fix (post-it overflow wrapping): hotfix new solution for overflow wrapping.
The last solution tried to  manually address the inconvenient cases. But not only wide letters cause trouble, but also very thin ones that would cause to break words on the same line. Also, differences in size between large and small post-its needed to be addressed. Earlier solution was discarded in favor of use of overflow-wrap property. This works very consistently. Drawback is there's no hyphen to breakwords. Also, this doesn't address vertical overflow (this was addressed with overflow-y property, though it's a suboptimal).